requirements.txt
# Generate requirements.txt with actually used packages with pipreqs . --force
python-dotenv
pydantic
pydantic-settings
pytest
pipreqs